Reorganize and streamline tasks
