+++ src/engine/combat.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-import { ActionDefaults, CombatStrategy as BaseCombatStrategy } from "grimoire-kolmafia";
+import { ActionDefaults, CombatStrategy } from "grimoire-kolmafia";
 import { $item, $skill, Macro } from "libram";
 
 const myActions = ["kill", "banish"] as const;
 export type CombatActions = typeof myActions[number];
-export class CombatStrategy extends BaseCombatStrategy.withActions(myActions) {}
+export class BaggoCombatStrategy extends CombatStrategy.withActions(myActions) {}
 export class MyActionDefaults implements ActionDefaults<CombatActions> {
   kill() {
     return this.delevel()
