+++ src/engine/task.ts
@@ -1,7 +1,7 @@
-import { CombatActions, CombatStrategy } from "./combat";
-import { Quest as BaseQuest, Task as BaseTask } from "grimoire-kolmafia";
+import { BaggoCombatStrategy, CombatActions } from "./combat";
+import { Quest, Task } from "grimoire-kolmafia";
 
-export type Quest = BaseQuest<Task>;
-export type Task = {
-  combat?: CombatStrategy;
-} & BaseTask<CombatActions>;
+export type BaggoQuest = Quest<BaggoTask>;
+export type BaggoTask = {
+  combat?: BaggoCombatStrategy;
+} & Task<CombatActions>;
